+++ test/122_gendata_h48short/gendata_h48short_tests.c
@@ -34,7 +34,7 @@ void run(void) {
 
 	capacity = readl();
 	randomizer = readl();
-	arg.maxdepth = readl();
+	arg.maxdepth = (uint8_t)readl();
 	arg.crep = crep;
 	arg.selfsim = selfsim;
 	arg.map = &map;
